+++ src/State.php
@@ -40,6 +40,10 @@ final class State
 
     public function __construct(Activity $activity, Actor $actor, $stateId, $registrationId = null)
     {
+        if (!$actor instanceof Agent) {
+            @trigger_error(sprintf('Passing a "%s" instance as the 2nd argument is deprecated since 1.2. In 3.0, only instance of "Xabbuh\XApi\Model\Agent" will be accepted.', get_class($actor)), E_USER_DEPRECATED);
+        }
+
         $this->activity = $activity;
         $this->actor = $actor;
         $this->stateId = $stateId;
@@ -60,8 +64,22 @@ final class State
      * Returns the actor.
      *
      * @return Actor The actor
+     *
+     * @deprecated since 1.2, to be removed in 3.0
      */
     public function getActor()
+    {
+        @trigger_error(sprintf('The "%s()" method is deprecated since 1.2, and will be removed in 3.0. Use "%s::getAgent()" method instead.', __METHOD__, __CLASS__), E_USER_DEPRECATED);
+
+        return $this->getAgent();
+    }
+
+    /**
+     * Returns the agent.
+     *
+     * @return Actor The agent
+     */
+    public function getAgent()
     {
         return $this->actor;
     }